packages/local/pyproject.toml
[project]
name = "sf6-chapter-local"
version = "0.1.0"
description = "SF6 YouTube動画処理ローカルスクリプト"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-pubsub>=2.18.0",
    "yt-dlp>=2024.0.0",
    "opencv-python>=4.8.0",
    "google-generativeai>=0.3.0",
    "google-api-python-client>=2.100.0",
    "boto3>=1.28.0",
    "pyarrow>=14.0.0",
    "jsonschema>=4.20.0",
    "pillow>=12.0.0",
    "google-auth-oauthlib>=1.2.3",
    "jupyter>=1.1.1",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
    "E402", # module level import not at top of file (sys.path manipulation required)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
